Rename `client` module, and move default argument from `handle` to `main`
