+++ src/fractal_client/parser/_main.py
@@ -28,7 +28,16 @@ def get_main_parser() -> ap.ArgumentParser:
     parser_main.add_argument(
         "--fractal-server",
         help=(
-            "URL of Fractal server (overrides `FRACTAL_SERVER` environment variable)."
+            "URL of Fractal server (overrides `FRACTAL_SERVER` environment variable). "
+            "Example: https://fractal.example.org/backend"
+        ),
+    )
+
+    parser_main.add_argument(
+        "--fractal-web",
+        help=(
+            "URL of Fractal web (overrides `FRACTAL_WEB` environment variable)."
+            "Example: https://fractal.example.org"
         ),
     )
