+++ tests/test_main.py
@@ -1,7 +1,7 @@
 import pytest
 
-from fractal_client.client import main
 from fractal_client.interface import Interface
+from fractal_client.main import main
 
 
 def test_unit_main(monkeypatch):
@@ -11,12 +11,12 @@ def test_unit_main(monkeypatch):
     NOTE: Mocking `handle()` is necessary because there is no
     appropriate `sys.argv`.
     """
-    import fractal_client.client
+    import fractal_client.main
 
     monkeypatch.setattr(
-        fractal_client.client,
+        fractal_client.main,
         "handle",
-        lambda: Interface(data="data", retcode=0),
+        lambda _: Interface(data="data", retcode=0),
     )
     with pytest.raises(SystemExit):
         main()
